backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20

# Security and rate limiting
slowapi>=0.1.9

# Vector database and embeddings
chromadb>=0.5.0
fastembed>=0.7.0

# PDF processing
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.0
Pillow>=11.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.67.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Note: Switched from sentence-transformers to fastembed for lower memory usage
# fastembed uses ONNX runtime instead of PyTorch (~80% less memory)
